final_project/house_rental_system/requirements.txt
asgiref==3.6.0
Django==4.1.4
django-phonenumber-field==7.0.1
phonenumberslite==8.13.3
sqlparse==0.4.3
tzdata==2022.7